aircraft-maintenance-sim/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "aircraft-maintenance-sim"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "networkx>=3.3",
    "pandas>=2.2",
    "pyarrow>=16.0",
    "numpy>=1.26",
    "streamlit>=1.35",
    "rich>=13.7",
    "python-dotenv>=1.0",
    "requests>=2.32",
    "scipy>=1.13",
    "pydantic>=2.7",
    "matplotlib>=3.9",
]

[project.optional-dependencies]
dev = ["pytest>=8.2", "pytest-cov", "ruff", "mypy"]

[tool.setuptools.packages.find]
where = ["src"]
